+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 0.3.0 - November 5, 2015
 ------------------------
 
+* Add ``overwrite`` flag to :meth:`.TableSQL.to_sql`.
 * Removed Python 2.6 support.
 * Updated agate dependency to version 1.1.0.
 * Additional SQL types are now supported. (#4, #10)